feat(tests): extract and reuse MockInMemoryDataGateway in domain and repository tests
